python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compresr"
version = "2.0.0"
description = "Python SDK for Compresr - Intelligent prompt compression service"
readme = "README.md"
license = "LicenseRef-Proprietary"
requires-python = ">=3.9"
authors = [
    {name = "Compresr Team", email = "founders@compresr.ai"}
]
keywords = ["llm", "compression", "ai", "openai", "gpt", "tokens", "cost-optimization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://compresr.ai"
Documentation = "https://compresr.ai/docs"
Repository = "https://github.com/compresr/sdk"
Issues = "https://github.com/compresr/sdk/issues"

[tool.setuptools]
packages = ["compresr", "compresr.services", "compresr.schemas", "compresr.exceptions"]

[tool.setuptools.package-data]
compresr = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
